fix: instead of `--part.userParticleHandler=''`, reduce `part.minimalKineticEnergy`
So far we run `npsim` with `--part.userParticleHandler=''`, which we
thought we needed in order to see optical photons in the truth output.
Today I learned we don't see them because the default `part.minimalKineticEnergy`
is 1 MeV, too high for photons. Instead, we set this minimal energy to
zero, and check if IRT reconstruction still works.